requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=1.4.0=pyhd8ed1ab_0
aom=3.5.0=h27087fc_0
astunparse=1.6.3=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
brotlipy=0.7.0=py37h540881e_1004
bzip2=1.0.8=h7f98852_4
c-ares=1.19.0=hd590300_0
ca-certificates=2023.5.7=hbcca054_0
cachetools=5.3.0=pypi_0
cairo=1.16.0=h35add3b_1015
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py37h43b0acd_1
charset-normalizer=3.1.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=3.4.8=py37hf1a17b8_1
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cuda-version=12.0=hffde075_2
cycler=0.11.0=pypi_0
cython=0.29.34=pypi_0
dav1d=1.2.0=hd590300_0
debugpy=1.5.1=py37h295c915_0
decorator=4.4.2=pypi_0
dlib=19.24.1=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
ffmpeg=6.0.0=gpl_h5bb71f3_100
fftw=3.3.10=nompi_hc118613_107
flatbuffers=23.5.9=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=pypi_0
freetype=2.12.1=hca18f0e_1
fribidi=1.0.10=h36c2ea0_0
gast=0.4.0=pypi_0
gettext=0.21.1=h27087fc_0
gmp=6.2.1=h58526e2_0
gnutls=3.7.8=hf3e180e_0
google-auth=2.18.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.42.0=py37hce63b2e_0
h5py=3.8.0=pypi_0
harfbuzz=7.3.0=hdb3a94d_0
icu=72.1=hcb278e6_0
idna=3.4=pyhd8ed1ab_0
imageio=2.28.1=pypi_0
importlib-metadata=4.11.4=py37h89c1867_0
ipykernel=6.15.2=py37h06a4308_0
ipython=7.33.0=py37h89c1867_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=2.11.3=pyhd8ed1ab_2
joblib=1.2.0=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.11.2=py37h06a4308_0
keras=2.11.0=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h166bdaf_1003
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libass=0.17.1=hc9aadba_0
libclang=16.0.0=pypi_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.5.0.59=hcb278e6_0
libcurand=10.3.1.50=hcb278e6_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.14=h166bdaf_0
libdrm=2.4.114=h166bdaf_0
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=12.2.0=h65d4601_19
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libglib=2.76.2=hebfc3b9_0
libhwloc=2.9.1=hd6dc26d_0
libiconv=1.17=h166bdaf_0
libidn2=2.3.4=h166bdaf_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libopus=1.3.1=h7f98852_1
libpciaccess=0.17=h166bdaf_0
libpng=1.6.39=h753d276_0
libprotobuf=3.15.8=h780b84a_1
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.42.0=h2797004_0
libstdcxx-ng=12.2.0=h46fd767_19
libtasn1=4.19.0=h166bdaf_0
libtiff=4.4.0=h82bc61c_5
libunistring=0.9.10=h7f98852_0
libuuid=2.38.1=h0b41bf4_0
libva=2.18.0=h0b41bf4_0
libvpx=1.13.0=hcb278e6_0
libwebp-base=1.3.0=h0b41bf4_0
libxcb=1.13=h7f98852_1004
libxml2=2.10.4=hfdac1af_0
libzlib=1.2.13=h166bdaf_4
llvm-openmp=16.0.4=h4dfa4b3_0
markdown=3.4.3=pyhd8ed1ab_0
markupsafe=1.1.1=py37h540881e_4
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mediapipe=0.9.0.1=pypi_0
mkl=2021.4.0=h8d4b97c_729
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h219a48f_0
mlxtend=0.22.0=pypi_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.8.1=hc379101_1
networkx=2.6.3=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
opencv-contrib-python=4.7.0.72=pypi_0
opencv-python=4.7.0.72=pypi_0
openh264=2.3.1=hcb278e6_2
openjpeg=2.5.0=h7d73246_1
openssl=1.1.1t=h0b41bf4_0
opt-einsum=3.3.0=pypi_0
p11-kit=0.24.1=hc5aa10d_0
packaging=23.1=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre2=10.40=hc3806b6_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.2.0=py37h850a105_2
pip=23.1.2=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
prompt-toolkit=3.0.38=pyha770c72_0
protobuf=3.15.8=py37hcd2ae1e_0
psutil=5.9.0=py37h5eee18b_0
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyg=2.3.0=py37_torch_1.13.0_cu117
pyg-lib=0.2.0+pt113cu117=pypi_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py37h89c1867_5
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.13.1=py3.7_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py37hac76be4_2
readline=8.2=h8228510_1
requests=2.29.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.42.0=h2c6b66d_0
svt-av1=1.5.0=h59595ed_0
tbb=2021.9.0=hf52228f_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorboardx=2.6=pypi_0
tensorflow=2.11.0=pypi_0
tensorflow-estimator=2.11.0=pypi_0
tensorflow-io-gcs-filesystem=0.32.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tifffile=2021.11.2=pypi_0
tk=8.6.12=h27826a3_0
torch-cluster=1.6.1+pt113cu117=pypi_0
torch-geometric-temporal=0.54.0=pypi_0
torch-scatter=2.1.1+pt113cu117=pypi_0
torch-sparse=0.6.17+pt113cu117=pypi_0
torch-spline-conv=1.2.2+pt113cu117=pypi_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=0.13.1=py37_cu117
torchvision=0.14.1=py37_cu117
tornado=6.2=py37h540881e_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
typing_extensions=4.5.0=pyha770c72_0
urllib3=1.26.15=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.40.0=pyhd8ed1ab_0
wrapt=1.15.0=pypi_0
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xorg-fixesproto=5.0=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.8.4=h0b41bf4_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.2=h5eee18b_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h166bdaf_4
zstd=1.5.2=h3eb15da_6
